pyproject.toml
[tool.poetry]
name = "negotia"
version = "0.1.0"
description = "Negotiate between llms"
authors = ["Mariano Falcon <mf2286@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
openai = "^1.6.1"
anthropic = "^0.39.0"
google-generativeai = "^0.8.3"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
jinja2 = "^3.1.6"
pyyaml = "^6.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
